fastapi/requirements.txt
# FastAPI framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# MongoDB async driver
motor>=3.3.0

# Environment variables
python-dotenv>=1.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Form data handling (required for OAuth2PasswordRequestForm)
python-multipart>=0.0.6

# Validation (included with FastAPI, but explicit)
pydantic>=2.0.0
email-validator>=2.0.0

# HTTP client for Colab communication
httpx>=0.27.0

# Evaluation metrics
evaluate>=0.4.0
pyvi>=0.1.1
bert-score>=0.3.13

# CSV/Excel processing
pandas>=2.0.0
openpyxl>=3.1.0
